Add routing with react-router
Create Home element which holds content of App. App imports Home and
sets it to be the element of the / path. BrowserRouter added toindex.js
